@@ -58,9 +58,7 @@ def check_args_not_none(*args, msg: str = None):
 def check_isinstance(item: T, classinfo: _ClassInfo) -> T:
     if not isinstance(item, classinfo):
         raise TypeError(
-            "Expected instance of type {!r} but got type {!r} for {!r}".format(
-                classinfo, type(item), item
-            )
+            f"Expected instance of type {classinfo} but got type {type(item)} for {item}"
         )
     return item
 
@@ -73,9 +71,7 @@ def check_opt_isinstance(item: T, classinfo: _ClassInfo) -> T:
     """
     if item and not isinstance(item, classinfo):
         raise TypeError(
-            "Expected instance of type {!r} but got type {!r} for {!r}".format(
-                classinfo, type(item), item
-            )
+            f"Expected instance of type {classinfo} but got type {type(item)} for {item}"
         )
     return item
 
@@ -87,9 +83,7 @@ def check_all_isinstance(items: Iterable[Any], classinfo: _ClassInfo):
 
 def check_issubclass(item, classinfo: _ClassInfo):
     if not issubclass(item, classinfo):
-        raise TypeError(
-            "Expected subclass of type {!r} but got {!r}".format(classinfo, type(item))
-        )
+        raise TypeError(f"Expected subclass of type {classinfo} but got {item}")
     return item
 
 
@@ -100,9 +94,8 @@ def check_in(item: Any, legal_values: Iterable[Any], item_name: str = None) -> N
         # which will presumably crash the program
         import vistautils.misc_utils  # pylint:disable=import-outside-toplevel
 
-        item_msg = " {!s} to be".format(item_name) if item_name else ""
+        item_msg = f" {item_name} to be" if item_name else ""
         raise ValueError(
-            "Expected{!s} one of {!s} but got {!s}".format(
-                item_msg, vistautils.misc_utils.str_list_limited(legal_values, 10), item
-            )
+            f"Expected{item_msg} one of {vistautils.misc_utils.str_list_limited(legal_values, 10)}"
+            f"but got {item}",
         )
